Add a `date` subcommand…
… and print errors in os.Stderr instead of stdout
