+++ Makefile
@@ -2,10 +2,20 @@ depends:
 	dep ensure
 
 test:
-	go test -v .
+	go test -v ./...
 
-build: test
-	rm -rf build
-	mkdir build
-	GOOS=darwin GOARCH=amd64 go build -o build/stupid
-	GOOS=windows GOARCH=amd64 go build -o build/stupid.exe
+bin/%: cmd/%
+	go build -o $@ ./$<
+
+bin/stupid-linux: cmd/stupid
+	GOOS=linux go build -o $@ ./$<
+
+bin/stupid-darwin: cmd/stupid
+	GOOS=darwin go build -o $@ ./$<
+
+bin/stupid-windows.exe: cmd/stupid
+	GOOS=windows go build -o $@ ./$<
+
+cross: bin/stupid-linux bin/stupid-darwin bin/stupid-windows.exe
+
+build: test cross
