+++ cmd/stupid/copy_test.go
@@ -9,61 +9,6 @@ import (
 	"gotest.tools/fs"
 )
 
-func TestRemove(t *testing.T) {
-	rootDirectory := fs.NewDir(t, "root",
-		fs.WithDir("empty-dir"),
-		fs.WithDir("full-dir",
-			fs.WithFile("some-file", "")),
-		fs.WithFile("file", ""),
-		fs.WithFile("remaining-file", ""),
-		fs.WithDir("remaining-dir"))
-	defer rootDirectory.Remove()
-
-	err := remove([]string{
-		filepath.Join(rootDirectory.Path(), "empty-dir"),
-		filepath.Join(rootDirectory.Path(), "full-dir"),
-		filepath.Join(rootDirectory.Path(), "non-existing"),
-		filepath.Join(rootDirectory.Path(), "file"),
-	})
-	assert.NilError(t, err)
-
-	expected := fs.Expected(t,
-		fs.WithFile("remaining-file", ""),
-		fs.WithDir("remaining-dir"))
-	assert.Assert(t, fs.Equal(rootDirectory.Path(), expected))
-}
-
-func TestRemoveWithGlob(t *testing.T) {
-	rootDirectory := fs.NewDir(t, "root",
-		fs.WithDir("empty-dir"),
-		fs.WithDir("full-dir",
-			fs.WithFile("some-file", "")),
-		fs.WithFile("remaining-file", ""))
-	defer rootDirectory.Remove()
-
-	err := remove([]string{
-		filepath.Join(rootDirectory.Path(), "*-dir"),
-		filepath.Join(rootDirectory.Path(), "full-dir"),
-	})
-	assert.NilError(t, err)
-
-	expected := fs.Expected(t, fs.WithFile("remaining-file", ""))
-	assert.Assert(t, fs.Equal(rootDirectory.Path(), expected))
-}
-
-func TestRemoveWithNonExistingGlob(t *testing.T) {
-	rootDirectory := fs.NewDir(t, "root",
-		fs.WithFile("remaining-file", ""))
-	defer rootDirectory.Remove()
-
-	err := remove([]string{
-		filepath.Join(rootDirectory.Path(), "non-existing*"),
-	})
-	assert.NilError(t, err)
-	expected := fs.Expected(t, fs.WithFile("remaining-file", ""))
-	assert.Assert(t, fs.Equal(rootDirectory.Path(), expected))
-}
-
 func TestCopyFileToNonExistingFile(t *testing.T) {
 	rootDirectory := fs.NewDir(t, "root",
 		fs.WithFile("foo.txt", "foo"))
@@ -267,100 +212,5 @@ func TestCopyTreeWithEmptyGlob(t *testing.T) {
 	expected := fs.Expected(t,
 		fs.WithDir("source",
 			fs.WithFile("foo.txt", "foo")))
-	assert.Assert(t, fs.Equal(rootDirectory.Path(), expected))
-}
-
-func TestTarTree(t *testing.T) {
-	rootDirectory := fs.NewDir(t, "root",
-		fs.WithDir("source",
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))))
-	defer rootDirectory.Remove()
-
-	dst := filepath.Join(rootDirectory.Path(), "destination", "dst.tar")
-	err := tarFiles(dst, filepath.Join(rootDirectory.Path(), "source", "foo.txt"), filepath.Join(rootDirectory.Path(), "source", "bar"))
-	assert.NilError(t, err)
-	err = untar(dst, filepath.Join(rootDirectory.Path(), "destination"))
-	assert.NilError(t, err)
-
-	expected := fs.Expected(t,
-		fs.WithDir("source",
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))),
-		fs.WithDir("destination",
-			fs.WithFile("dst.tar", "", fs.MatchAnyFileContent),
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))))
-	assert.Assert(t, fs.Equal(rootDirectory.Path(), expected))
-}
-
-func TestTarTreeWithGlob(t *testing.T) {
-	rootDirectory := fs.NewDir(t, "root",
-		fs.WithDir("source",
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))))
-	defer rootDirectory.Remove()
-
-	dst := filepath.Join(rootDirectory.Path(), "destination", "dst.tar")
-	err := tarFiles(dst, filepath.Join(rootDirectory.Path(), "source", "*"))
-	assert.NilError(t, err)
-	err = untar(dst, filepath.Join(rootDirectory.Path(), "destination"))
-	assert.NilError(t, err)
-
-	expected := fs.Expected(t,
-		fs.WithDir("source",
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))),
-		fs.WithDir("destination",
-			fs.WithFile("dst.tar", "", fs.MatchAnyFileContent),
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))))
-	assert.Assert(t, fs.Equal(rootDirectory.Path(), expected))
-}
-
-func TestTarTreeWithEmptyGlob(t *testing.T) {
-	rootDirectory := fs.NewDir(t, "root",
-		fs.WithDir("source",
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))))
-	defer rootDirectory.Remove()
-
-	src := filepath.Join(rootDirectory.Path(), "source", "non-existing*")
-	dst := filepath.Join(rootDirectory.Path(), "destination", "dst.tar")
-	err := tarFiles(dst, src)
-	assert.Error(t, err, "Source ["+src+"] does not exist")
-}
-
-func TestGzipTarTree(t *testing.T) {
-	rootDirectory := fs.NewDir(t, "root",
-		fs.WithDir("source",
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))))
-	defer rootDirectory.Remove()
-
-	dst := filepath.Join(rootDirectory.Path(), "destination", "dst.tar.gz")
-	err := tarFiles(dst, filepath.Join(rootDirectory.Path(), "source", "foo.txt"), filepath.Join(rootDirectory.Path(), "source", "bar"))
-	assert.NilError(t, err)
-	err = untar(dst, filepath.Join(rootDirectory.Path(), "destination"))
-	assert.NilError(t, err)
-
-	expected := fs.Expected(t,
-		fs.WithDir("source",
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))),
-		fs.WithDir("destination",
-			fs.WithFile("dst.tar.gz", "", fs.MatchAnyFileContent),
-			fs.WithFile("foo.txt", "foo\n"),
-			fs.WithDir("bar",
-				fs.WithFile("bar.txt", "bar\n"))))
 	assert.Assert(t, fs.Equal(rootDirectory.Path(), expected))
 }
